initial work towards updating menus with virtual DOM.
